refactor: Revise methods
Alter `Hierarchy` methods to return a `Hierarchy` & `List` rather than a
`[]T`, former two can be converted to the latter by their respective
`...Values` methods. This brings consistency to the availed methods.

Move `lexer.Lexer` instantiation outside of `Deserialize`. This allows
for configuration of the lexer and the generated `Hierarchy`; prior
implementation configured the lexer only.